not using Docker env_file
We want to escape from hell. On the one hand, Docker needs the .env file for the
env_file configuration.  On the other hand, .env can't be checked into Git as an
empty file, because then changes will be tracked.

Solution is to source the .env file on Rails startup.